server/requirements.txt
faiss-cpu==1.8.0
langchain==0.1.19
langchain-anthropic==0.1.11
langchain-community==0.0.38
langchain-core==0.1.52
langchain-experimental==0.0.58
langchain-openai==0.1.6
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langsmith==0.1.56
langgraph
langserve
langchain-pinecone  # Added langchain-pinecone
pinecone-notebooks  # Added pinecone-notebooks
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.8.4
pandas==1.5.3
pandasai==2.0.37
protobuf==4.25.3
pyarrow==16.0.0
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.0
PyMuPDF==1.24.4
PyMuPDFb==1.24.3
pyparsing==3.1.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.4.28
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.18.1
scipy==1.13.0
smmap==5.0.1
sniffio==1.3.1
specklepy==2.18.3
SQLAlchemy==2.0.30
streamlit==1.34.0
stringcase==1.2.0
tabulate==0.9.0
tenacity==8.3.0
tiktoken==0.6.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
tqdm==4.66.4
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
websockets==11.0.3
wrapt==1.16.0
yarl==1.9.4
ipykernel
document
firecrawl-py
fastapi==0.110.2
uvicorn==0.29.0
sse_starlette
gradio
